refactor: centralize script conditions from state modules
Move all script conditions defined in state modules into the
`conditions` module.

Leave the `Variables` module alone as the script conditions in it
seem to be unused and misnamed.
